+++ chromium_src/chrome/browser/ui/tabs/tab_strip_model.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2019 The Brave Authors. All rights reserved.
+/* Copyright (c) 2020 The Brave Authors. All rights reserved.
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
@@ -11,5 +11,6 @@
 
 #include "../../../../../../chrome/browser/ui/tabs/tab_strip_model.h"  // NOLINT
 #undef SelectRelativeTab
+#undef TAB_STRIP_MODEL_H_
 
 #endif  // BRAVE_CHROMIUM_SRC_CHROME_BROWSER_UI_TABS_TAB_STRIP_MODEL_H_
